code/app/web/requirements.txt
pyserial==3.4
Flask==1.0.2
Flask-SocketIO==3.0.2
numpy==1.16.4
gevent==1.4.0
greenlet==0.4.15